+++ src/main/java/com/cburch/logisim/util/AutoLabel.java
@@ -19,6 +19,8 @@ import com.cburch.logisim.data.AttributeSet;
 import com.cburch.logisim.fpga.designrulecheck.CorrectLabel;
 import com.cburch.logisim.gui.generic.OptionPane;
 import com.cburch.logisim.instance.StdAttr;
+import com.cburch.logisim.prefs.AppPreferences;
+import com.cburch.logisim.prefs.PrefMonitorKeyStroke;
 import com.cburch.logisim.std.wiring.Tunnel;
 import com.cburch.logisim.tools.SetAttributeAction;
 import java.awt.event.KeyEvent;
@@ -31,7 +33,8 @@ import java.util.TreeSet;
 
 public class AutoLabel {
 
-  public static final Integer[] USED_KEY_STROKES = new Integer[] {KeyEvent.VK_L, KeyEvent.VK_T, KeyEvent.VK_V, KeyEvent.VK_H, KeyEvent.VK_A};
+  public static final Integer[] USED_KEY_STROKES = new Integer[] {
+      KeyEvent.VK_L, KeyEvent.VK_T, KeyEvent.VK_V, KeyEvent.VK_H, KeyEvent.VK_A};
   public static final Set<Integer> KEY_STROKES = new HashSet<>(Arrays.asList(USED_KEY_STROKES));
 
   private final HashMap<Circuit, String> labelBase = new HashMap<>();
@@ -55,7 +58,9 @@ public class AutoLabel {
   }
 
   public boolean hasNext(Circuit circ) {
-    if (circ == null || !active.containsKey(circ)) return false;
+    if (circ == null || !active.containsKey(circ)) {
+      return false;
+    }
     return active.get(circ);
   }
 
@@ -63,7 +68,8 @@ public class AutoLabel {
     if (circ == null || !currentLabel.containsKey(circ) || currentLabel.get(circ).isEmpty()) {
       return "";
     }
-    if (Circuit.isCorrectLabel(circ.getName(), currentLabel.get(circ), circ.getNonWires(), null, me, false)) {
+    if (Circuit.isCorrectLabel(circ.getName(),
+        currentLabel.get(circ), circ.getNonWires(), null, me, false)) {
       return currentLabel.get(circ);
     }
 
@@ -75,29 +81,43 @@ public class AutoLabel {
     return "";
   }
 
-  public boolean correctMatrixBaseLabel(Circuit circ, ComponentFactory me, String common, int maxX, int maxY) {
-    if (StringUtil.isNullOrEmpty(common) || (maxX < 0) || (maxY < 0)) return true;
-    if (!SyntaxChecker.isVariableNameAcceptable(common, true)) return false;
-    for (var x = 0; x < maxX; x++)
+  public boolean correctMatrixBaseLabel(Circuit circ, ComponentFactory me,
+                                        String common, int maxX, int maxY) {
+    if (StringUtil.isNullOrEmpty(common) || (maxX < 0) || (maxY < 0)) {
+      return true;
+    }
+    if (!SyntaxChecker.isVariableNameAcceptable(common, true)) {
+      return false;
+    }
+    for (var x = 0; x < maxX; x++) {
       for (var y = 0; y < maxY; y++) {
         if (getMatrixLabel(circ, me, common, x, y).isEmpty()) {
           return false;
         }
       }
+    }
     return true;
   }
 
   public String getMatrixLabel(Circuit circ, ComponentFactory me, String common, int x, int y) {
-    if (StringUtil.isNullOrEmpty(common) || (x < 0) || (y < 0)) return "";
-    if (circ == null || !currentLabel.containsKey(circ) || currentLabel.get(circ).isEmpty()) return "";
+    if (StringUtil.isNullOrEmpty(common) || (x < 0) || (y < 0)) {
+      return "";
+    }
+    if (circ == null || !currentLabel.containsKey(circ) || currentLabel.get(circ).isEmpty()) {
+      return "";
+    }
     final var label = common.concat("_X" + x + "_Y" + y);
     if (Circuit.isCorrectLabel(circ.getName(), label, circ.getNonWires(), null, me, false)
-        && SyntaxChecker.isVariableNameAcceptable(label, false)) return label;
+        && SyntaxChecker.isVariableNameAcceptable(label, false)) {
+      return label;
+    }
     return "";
   }
 
   public String getNext(Circuit circ, ComponentFactory me) {
-    if (circ == null) return "";
+    if (circ == null) {
+      return "";
+    }
     if (useLabelBaseOnly.get(circ)) {
       useLabelBaseOnly.put(circ, false);
       return labelBase.get(circ);
@@ -112,21 +132,28 @@ public class AutoLabel {
     do {
       curIdx++;
       newLabel = baseLabel;
-      if (undescore) newLabel = newLabel.concat("_");
+      if (undescore) {
+        newLabel = newLabel.concat("_");
+      }
       newLabel = newLabel.concat(Integer.toString(curIdx));
-    } while (!Circuit.isCorrectLabel(circ.getName(), newLabel, circ.getNonWires(), null, me, false));
+    } while (!Circuit.isCorrectLabel(circ.getName(),
+        newLabel, circ.getNonWires(), null, me, false));
     currentIndex.put(circ, curIdx);
     currentLabel.put(circ, newLabel);
     return newLabel;
   }
 
   public boolean isActive(Circuit circ) {
-    if (circ != null && active.containsKey(circ)) return active.get(circ);
+    if (circ != null && active.containsKey(circ)) {
+      return active.get(circ);
+    }
     return false;
   }
 
   public void setLabel(String label, Circuit circ, ComponentFactory me) {
-    if (circ != null) update(circ, label, true, me);
+    if (circ != null) {
+      update(circ, label, true, me);
+    }
   }
 
   public void activate(Circuit circ) {
@@ -134,7 +161,9 @@ public class AutoLabel {
       if (labelBase.containsKey(circ)
           && currentIndex.containsKey(circ)
           && useLabelBaseOnly.containsKey(circ)
-          && useUnderscore.containsKey(circ)) active.put(circ, !labelBase.get(circ).isEmpty());
+          && useUnderscore.containsKey(circ)) {
+        active.put(circ, !labelBase.get(circ).isEmpty());
+      }
     }
   }
 
@@ -151,12 +180,16 @@ public class AutoLabel {
 
   private int getLabelBaseEndIndex(String label) {
     var index = label.length();
-    while ((index > 1) && CorrectLabel.NUMBERS.contains(label.substring(index - 1, index))) index--;
+    while ((index > 1) && CorrectLabel.NUMBERS.contains(label.substring(index - 1, index))) {
+      index--;
+    }
     return (index - 1);
   }
 
   private void update(Circuit circ, String label, boolean useFirstLabel, ComponentFactory me) {
-    if (circ == null) return;
+    if (circ == null) {
+      return;
+    }
     if (label.isEmpty() || !SyntaxChecker.isVariableNameAcceptable(label, false)) {
       labelBase.put(circ, "");
       currentIndex.put(circ, 0);
@@ -189,20 +222,25 @@ public class AutoLabel {
     return sorted;
   }
 
-  public String askAndSetLabel(String componentName, String oldLabel, Circuit circ, Component comp, ComponentFactory compFactory,
-      AttributeSet attrs, SetAttributeAction act, boolean createAction) {
+  public String askAndSetLabel(String componentName, String oldLabel, Circuit circ,
+                               Component comp, ComponentFactory compFactory,
+                               AttributeSet attrs, SetAttributeAction act, boolean createAction) {
     var correct = false;
     var newLabel = oldLabel;
     while (!correct) {
       newLabel = (String)
-              OptionPane.showInputDialog(null, S.get("editLabelQuestion") + " " + componentName,
-                  S.get("editLabelDialog"), OptionPane.QUESTION_MESSAGE, null, null, oldLabel);
+          OptionPane.showInputDialog(null, S.get("editLabelQuestion") + " " + componentName,
+              S.get("editLabelDialog"), OptionPane.QUESTION_MESSAGE, null, null, oldLabel);
       if (newLabel != null) {
-        if (Circuit.isCorrectLabel(circ.getName(), newLabel, circ.getNonWires(), attrs, compFactory, true)
+        if (Circuit.isCorrectLabel(circ.getName(),
+            newLabel, circ.getNonWires(), attrs, compFactory, true)
             && SyntaxChecker.isVariableNameAcceptable(newLabel, true)
             && !CorrectLabel.isKeyword(newLabel, true)) {
-          if (createAction) act.set(comp, StdAttr.LABEL, newLabel);
-          else setLabel(newLabel, circ, compFactory);
+          if (createAction) {
+            act.set(comp, StdAttr.LABEL, newLabel);
+          } else {
+            setLabel(newLabel, circ, compFactory);
+          }
           correct = true;
         }
       } else {
@@ -213,60 +251,63 @@ public class AutoLabel {
     return newLabel;
   }
 
-  public boolean labelKeyboardHandler(int keyCode, AttributeSet attrs, String componentName, Component comp,
-      ComponentFactory compFactory, Circuit circ, SetAttributeAction act, boolean createAction) {
-    switch (keyCode) {
-      case KeyEvent.VK_L -> {
-        if (attrs.containsAttribute(StdAttr.LABEL)) {
-          final var oldLabel = attrs.getValue(StdAttr.LABEL);
-          final var newLabel = askAndSetLabel(componentName, oldLabel, circ, comp, compFactory,
-              attrs, act, createAction);
-          if (!newLabel.equals(oldLabel)) {
-            if (!newLabel.isEmpty() && labelEndsWithNumber(newLabel)) {
-              activate(circ);
-            } else {
-              active.put(circ, false);
-            }
+  public boolean labelKeyboardHandler(int keyCode, int keyModifiers, AttributeSet attrs,
+                                      String componentName, Component comp,
+                                      ComponentFactory compFactory, Circuit circ,
+                                      SetAttributeAction act, boolean createAction) {
+    /* TODO: bind more hotkeys */
+    int code = keyCode;
+    int modifier = keyModifiers;
+    if (((PrefMonitorKeyStroke) AppPreferences.HOTKEY_AUTO_LABEL_OPEN).compare(code, modifier)) {
+      if (attrs.containsAttribute(StdAttr.LABEL)) {
+        final var oldLabel = attrs.getValue(StdAttr.LABEL);
+        final var newLabel = askAndSetLabel(componentName, oldLabel, circ, comp, compFactory,
+            attrs, act, createAction);
+        if (!newLabel.equals(oldLabel)) {
+          if (!newLabel.isEmpty() && labelEndsWithNumber(newLabel)) {
+            activate(circ);
+          } else {
+            active.put(circ, false);
           }
         }
-        return true;
       }
-      case KeyEvent.VK_T -> {
-        if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)) {
-          if (createAction)
-            act.set(comp, StdAttr.LABEL_VISIBILITY, !attrs.getValue(StdAttr.LABEL_VISIBILITY));
-          else
-            attrs.setValue(StdAttr.LABEL_VISIBILITY, !attrs.getValue(StdAttr.LABEL_VISIBILITY));
+      return true;
+    } else if (((PrefMonitorKeyStroke) AppPreferences.HOTKEY_AUTO_LABEL_TOGGLE)
+        .compare(code, modifier)) {
+      if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)) {
+        if (createAction) {
+          act.set(comp, StdAttr.LABEL_VISIBILITY, !attrs.getValue(StdAttr.LABEL_VISIBILITY));
+        } else {
+          attrs.setValue(StdAttr.LABEL_VISIBILITY, !attrs.getValue(StdAttr.LABEL_VISIBILITY));
         }
-        return true;
       }
-      case KeyEvent.VK_V -> {
-        if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)
-            && !attrs.getValue(StdAttr.LABEL_VISIBILITY)) {
-          if (createAction)
-            act.set(comp, StdAttr.LABEL_VISIBILITY, true);
-          else
-            attrs.setValue(StdAttr.LABEL_VISIBILITY, true);
+      return true;
+    } else if (((PrefMonitorKeyStroke) AppPreferences.HOTKEY_AUTO_LABEL_VIEW)
+        .compare(code, modifier)) {
+      if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)
+          && !attrs.getValue(StdAttr.LABEL_VISIBILITY)) {
+        if (createAction) {
+          act.set(comp, StdAttr.LABEL_VISIBILITY, true);
+        } else {
+          attrs.setValue(StdAttr.LABEL_VISIBILITY, true);
         }
-        return true;
       }
-      case KeyEvent.VK_H -> {
-        if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)
-            && attrs.getValue(StdAttr.LABEL_VISIBILITY)) {
-          if (createAction)
-            act.set(comp, StdAttr.LABEL_VISIBILITY, false);
-          else
-            attrs.setValue(StdAttr.LABEL_VISIBILITY, false);
+      return true;
+    } else if (((PrefMonitorKeyStroke) AppPreferences.HOTKEY_AUTO_LABEL_HIDE)
+        .compare(code, modifier)) {
+      if (attrs.containsAttribute(StdAttr.LABEL_VISIBILITY)
+          && attrs.getValue(StdAttr.LABEL_VISIBILITY)) {
+        if (createAction) {
+          act.set(comp, StdAttr.LABEL_VISIBILITY, false);
+        } else {
+          attrs.setValue(StdAttr.LABEL_VISIBILITY, false);
         }
-        return true;
       }
-      case KeyEvent.VK_A -> {
-        stop(circ);
-        return true;
-      }
-      default -> {
-        // nothing
-      }
+      return true;
+    } else if (((PrefMonitorKeyStroke) AppPreferences.HOTKEY_AUTO_LABEL_SELF_NUMBERED_STOP)
+        .compare(code, modifier)) {
+      stop(circ);
+      return true;
     }
     return false;
   }